fix board_column (#70)

+++ feature/board_columns/transport/delete_board_columns_handler.go
@@ -5,15 +5,16 @@ import (
 	"github.com/gofiber/fiber/v2"
 )
 
-// deleteBoardColumn godoc
+// deleteBoardColumn
 // @Summary Delete a board column (X-User-Email required, X-Workspace-Id required)
 // @Description Delete a board column (X-User-Email required, X-Workspace-Id required)
 // @Tags board_columns
 // @Accept json
 // @Produce json
 // @Param board_column_id path string true "Board column ID"
-// @Param email path string true "User email"
-// @Param workspace_id path string true "Workspace ID"
+// @Param X-User-Email header string true "User email"
+// @Param X-Workspace-Id header string true "Workspace ID"
+// @Security bearerToken
 // @Router /api/v1/board_columns/{board_column_id} [delete]
 func (h *BoardColumnsHandler) deleteBoardColumn(c *fiber.Ctx) error {
 	// Parse the request
@@ -23,15 +24,26 @@ func (h *BoardColumnsHandler) deleteBoardColumn(c *fiber.Ctx) error {
 			"message": "Board column id is required",
 		})
 	}
+	board_column, err1 := board_columns.NewBoardColumnsService().GetBoardColumnById(boardColumnId)
+	if err1 != nil {
+		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
+			"message": "Internal server error",
+		})
+	}
 
 	// Delete the board column
 	var err = board_columns.NewBoardColumnsService().DeleteBoardColumn(boardColumnId)
 	if err != nil {
 		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
+			"message": "Internal server error",
+		})
+	}
+	err = board_columns.NewBoardColumnsService().UpdatePositionsAfterDeletion(board_column.Position, board_column.WorkspaceId)
+	if err != nil {
+		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
 			"message": err.Error(),
 		})
 	}
-
 	// Return the response
 	return c.JSON(fiber.Map{
 		"message": "Board column deleted successfully",
